requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.8.5
python-dotenv==1.0.0
Pillow>=10.2.0
pydantic>=2.9.0
setuptools>=65.5.0